feat(learn): move learn page and add ClassroomSubjects
